Alhamd le Allah

+++ 0x1E-search_algorithms/0-linear.c
@@ -10,11 +10,11 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	size_t index = 0;
+	int index = 0;
 
 	if (array != NULL)
 	{
-		for (index = 0; index < size; index++)
+		for (index = 0; index < (int)size; index++)
 		{
 			printf("Value checked array[%d] = [%d]\n", index, array[index]);
 			if (value == array[index])
